+++ workspaces/redhat-argocd/plugins/argocd/src/components/DeploymentLifeCycle/sidebar/resources/resource/DeploymentMetadata.tsx
@@ -24,6 +24,7 @@ import MetadataItem from '../../../../Common/MetadataItem';
 import MetadataItemWithTooltip from '../../../../Common/MetadataItemWithTooltip';
 import AppCommitLink from '../../../../Common/AppCommitLink';
 import { DeploymentHistory } from './DeploymentHistory';
+import { useTranslation } from '../../../../../hooks/useTranslation';
 
 const useDeploymentInfoStyles = makeStyles((theme: Theme) => ({
   latestDeploymentContainer: {
@@ -73,14 +74,26 @@ const DeploymentMetadata = ({ resource }: { resource: Resource }) => {
       );
     });
   };
+
+  const { t } = useTranslation();
   return (
     <>
       <Metadata>
-        <MetadataItem title="Namespace">{resource?.namespace}</MetadataItem>
+        <MetadataItem
+          title={t(
+            'deploymentLifecycle.sidebar.resources.resource.deploymentMetadata.namespace',
+          )}
+        >
+          {resource?.namespace}
+        </MetadataItem>
         {appHistory.length > 0 ? (
           <MetadataItemWithTooltip
-            title="Image(s)"
-            tooltipText="These are the images for all the deployments in the ArgoCD application."
+            title={t(
+              'deploymentLifecycle.sidebar.resources.resource.deploymentMetadata.metadataItemWithTooltip.title',
+            )}
+            tooltipText={t(
+              'deploymentLifecycle.sidebar.resources.resource.deploymentMetadata.metadataItemWithTooltip.tooltipText',
+            )}
           >
             <ImageLinks />
           </MetadataItemWithTooltip>
@@ -88,7 +101,11 @@ const DeploymentMetadata = ({ resource }: { resource: Resource }) => {
           <></>
         )}
         {!isAppHelmChartType(application) ? (
-          <MetadataItem title="Commit">
+          <MetadataItem
+            title={t(
+              'deploymentLifecycle.sidebar.resources.resource.deploymentMetadata.commit',
+            )}
+          >
             <AppCommitLink
               application={application}
               entity={entity}