+++ workspaces/entity-feedback/plugins/entity-feedback/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage-community/plugin-entity-feedback
 
+## 0.11.1
+
+### Patch Changes
+
+- d5b74e9: Migrate the `entity-feedback` plugin to the new frontend system.
+
+## 0.11.0
+
+### Minor Changes
+
+- f652e7b: Backstage version bump to v1.43.2
+
+### Patch Changes
+
+- Updated dependencies [f652e7b]
+  - @backstage-community/plugin-entity-feedback-common@0.9.0
+
 ## 0.10.0
 
 ### Minor Changes