+++ workspaces/playlist/plugins/playlist-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-playlist-common
 
+## 0.10.0
+
+### Minor Changes
+
+- 6724458: Backstage version bump to v1.43.2
+
 ## 0.9.0
 
 ### Minor Changes