+++ workspaces/topology/plugins/topology/CHANGELOG.md
@@ -1,5 +1,17 @@
 ### Dependencies
 
+## 2.6.0
+
+### Minor Changes
+
+- 561e5d1: Add i18n support
+
+## 2.5.0
+
+### Minor Changes
+
+- c850e6d: Backstage version bump to v1.42.5
+
 ## 2.4.2
 
 ### Patch Changes